[L]Lab3/requirements.txt
psutil==6.0.0
inotify_simple==1.3.5
PyQt6==6.7.0
matplotlib==3.8.4
PyYAML==6.0.2
